requirements.txt
pyjwt==1.7.1
paho-mqtt==1.5.1
cryptography==3.2.1
google-cloud-iot
pycryptodome
Flask